requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
enum34==1.1.6
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.2
numpy-stl==2.7.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.5
six==1.11.0
subprocess32==3.5.3
